scripts/requirements.txt
markdownify==0.11.6
requests==2.28.1
python-slugify==6.1.2
pip==22.2.2
setuptools==65.3.0
pylint==2.15.0
beautifulsoup4==4.11.1
tabulate==0.8.10
lxml==4.9.1
